Separate domain models from API client
