+++ src/lifetime_bot/__init__.py
@@ -5,16 +5,16 @@ An automated bot for reserving classes at Life Time.
 
 __version__ = "1.0.0"
 
-from lifetime_bot.api import (
+from lifetime_bot.api import LifetimeAPIClient
+from lifetime_bot.bot import LifetimeReservationBot
+from lifetime_bot.config import BotConfig, EmailConfig, SMSConfig
+from lifetime_bot.errors import LifetimeAPIError
+from lifetime_bot.models import (
     ClassEvent,
-    LifetimeAPIClient,
-    LifetimeAPIError,
     RegistrationOutcome,
     RegistrationResult,
     SessionTokens,
 )
-from lifetime_bot.bot import LifetimeReservationBot
-from lifetime_bot.config import BotConfig, EmailConfig, SMSConfig
 
 __all__ = [
     "LifetimeReservationBot",
